+++ MegaManLofi/IArena.h
@@ -7,7 +7,7 @@ namespace MegaManLofi
    class __declspec( novtable ) IArena : public IArenaInfoProvider
    {
    public:
-      virtual void SetPlayerPositionX( double positionX ) = 0;
-      virtual void SetPlayerPositionY( double positionY ) = 0;
+      virtual void SetPlayerPositionX( long long positionX ) = 0;
+      virtual void SetPlayerPositionY( long long positionY ) = 0;
    };
 }
