+++ MegaManLofi/IArenaInfoProvider.h
@@ -7,14 +7,14 @@ namespace MegaManLofi
    class __declspec( novtable ) IArenaInfoProvider
    {
    public:
-      virtual double GetWidth() const = 0;
-      virtual double GetHeight() const = 0;
+      virtual long long GetWidth() const = 0;
+      virtual long long GetHeight() const = 0;
 
-      virtual double GetPlayerPositionX() const = 0;
-      virtual double GetPlayerPositionY() const = 0;
+      virtual long long GetPlayerPositionX() const = 0;
+      virtual long long GetPlayerPositionY() const = 0;
 
-      virtual double GetTileWidth() const = 0;
-      virtual double GetTileHeight() const = 0;
+      virtual long long GetTileWidth() const = 0;
+      virtual long long GetTileHeight() const = 0;
 
       virtual int GetHorizontalTiles() const = 0;
       virtual int GetVerticalTiles() const = 0;
